+++ qblox_drive_AS/Executions/m11_TimeRabiOsci.py
@@ -1,3 +1,4 @@
+""" This exp doesn't update any info for QDmanager """
 from qblox_drive_AS.support.Path_Book import find_latest_QD_pkl_for_dr
 from qblox_drive_AS.support import Data_manager
 from qblox_drive_AS.support.ExpFrames import TimeRabiOsci
